debugging and adding features
