+++ kod/printed_updated/210715-094622-esp12e/.vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "C_Cpp.errorSquiggles": "Disabled"
+}
