--- kod/printed_updated/210715-094622-esp12e/src/addnetio.html
@@ -1,55 +1,0 @@
-<html>
-<head>
-    <style>
-        button {
-            background-color: #005F41;
-            color: white;
-            margin: 10px;
-            padding: 15px 32px;
-            text-align: center;
-            font-size: 16px;
-            display: block;
-            border-radius: 4px;
-            width: 150px;
-            text-decoration: none;
-            transition: 0.3s;
-        }
-
-        button:hover {
-            background-color: #3d8b40;
-        }
-
-        input[type=text] {
-            width: 30%;
-            height: 40px;
-            margin: 10px;
-            font-size: large;
-        }
-
-        label {
-            font-size: 20px;
-        }
-
-        select {
-            margin: 20px;
-            font-size: 20px;
-        }
-    </style>
-
-</head>
-<body>
-<center>
-    <h1>Add Netio device</h1>
-    <form method="GET" action="/netioProduct/check">
-        <label for="addIP">IP of Netio socket</label> <br>
-        <input type="text" name="addIP" id="addIP"><br>
-        <label for="group">Button selection</label><br>
-        <select name="group" , id="group">
-            <option value=true>Button #1</option>
-            <option value=false>Button #2</option>
-        </select><br>
-        <button type="submit">Submit</button>
-    </form>
-</center>
-</body>
-</html>
