+++ kod/printed_updated/210715-094622-esp12e/src/define.h
@@ -1,0 +1,22 @@
+#ifndef _define_h_
+#define _define_h_
+
+
+// delky jsou vcetne 1 mezery pro ukonceni zapisu
+#define IP_POSA 0
+#define IP_POSB 188
+#define IP_LEN 187
+#define IP_JMP 16
+
+#define HTTP_POS1 400
+#define HTTP_POS2 601
+#define HTTP_LEN 200
+
+#define SSID_POS 802
+#define SSID_LEN 33
+
+#define PASSWORD_POS 836
+#define PASSWORD_LEN 100
+
+
+#endif
