--- kod/printed_updated/210715-094622-esp12e/src/ip_adress.json
@@ -1,1 +1,0 @@
-{"numOfIPA": "5", "IP_adressA": ["192.168.0.45", "192.168.0.45", "255.255.255.255", "255.255", "192.168.0.45" ], "numOfIPB": "1", "IP_addressB""192.168.0.45" }
