--- kod/printed_updated/210715-094622-esp12e/src/menu.html
@@ -1,131 +1,0 @@
-<html>
-<head>
-    <meta charset="UTF-8">
-    <style>
-        body {
-            font-family: "Helvetica";
-
-        }
-
-        input[type="submit"] {
-            background-color: #005F41; /* Green */
-            color: white;
-            width: 200px;
-            padding: 15px 32px;
-            text-align: center;
-            text-decoration: none;
-            display: inline-block;
-            font-size: 16px;
-            display: block;
-
-            margin: 0 auto;
-        }
-
-        button {
-            background-color: #005F41; /* Green */
-            color: white;
-            width: 150px;
-            padding: 15px 32px;
-            text-align: center;
-            text-decoration: none;
-            display: inline-block;
-            border-radius: 4px;
-            font-size: 16px;
-            display: block;
-            margin: 0 auto;
-            transition: 0.3s;
-        }
-
-        button:hover {
-            background-color: #3d8b40;
-        }
-
-        .rssi {
-            color: green;
-            text-align: center;
-        }
-
-        .encrypt1 {
-            color: green;
-            text-align: center;
-        }
-    </style>
-</head>
-<body>
-<center>
-    <h1>NETIO BUTTON</h1>
-    <button onclick="window.location.href=window.location.href">Scan</button>
-    <button onclick="location.href = '/netioProduct';">NETIO IP</button>
-    <button onclick="location.href = '/buttonConfigure';">Configure</button>
-    <button onclick="location.href = '/settings';">Settings</button>
-    <button onclick="location.href = '/deepsleep';">Deep sleep</button>
-    <button onclick="location.href = '/disconnect';">Disconnect</button>
-    <p>Scan can takes a while</p>
-    <h1>WIFIs</h1>
-
-    <script>
-        var getJSON = function (url, callback) {
-            var xhr = new XMLHttpRequest();
-            xhr.open('GET', url, true);  // Otevre http request
-            xhr.responseType = 'json';
-            xhr.onload = function () {
-                var status = xhr.status; // zkontroluje se status http
-                if (status === 200) {
-                    callback(null, xhr.response);
-                } else {
-                    callback(status, xhr.response);
-                }
-            };
-            xhr.send();
-        };
-        getJSON('scannedWiFi.json', // pracuje s ipAdresa/scannedWiFi.json
-            function (err, data) {
-                if (err !== null) { // kontrola zda se stal error
-                    alert('Something went wrong: ' + err);
-                } else { // vezme data z jsonu a ulozi se do pole
-                    for (i = 0; i < data.numOfNetworks; i++) {
-                        buttons(i, data.networks, data.strengh, data.protection);
-                    }
-                }
-            });
-
-        function buttons(i, networks, rssi, protection) {
-            // vytvori tlacitka na zaklade json zpravy
-            var button = document.createElement('input');
-            button.class = 'button';
-            button.type = 'submit';
-            button.value = networks[i]; // hodnota se nastavi na SSID site
-
-            var form = document.createElement('form'); // vytvori se formular do ktereho se tlacitko vlozi
-            // pokud sit je zabezpecena, odkaze na overeni s heslem
-            if (protection[i] == "None") {
-                form.method = 'POST';
-                form.action = '/wifi/redirect'
-            } else {
-                form.method = 'GET'
-                form.action = '/wifi';
-            }
-
-
-            var inputs = document.createElement('input');
-            inputs.type = 'hidden';
-            inputs.name = 'ssid';
-            inputs.value = networks[i]; // skryty input s ssid
-            var rssi1 = document.createElement("p");
-            rssi1.innerHTML = rssi[i]; // text s rrsi konkretni site
-            rssi1.className = 'rssi';
-            var encrypt = document.createElement("p");
-            encrypt.innerHTML = protection[i];
-            encrypt.className = 'encrypt1';
-            // sestaveni html stromu
-            document.body.appendChild(form);
-            form.appendChild(inputs);
-            form.appendChild(button);
-            form.appendChild(rssi1);
-            form.appendChild(encrypt);
-
-        }
-    </script>
-</center>
-</body>
-</html>
