+++ kod/printed_updated/210715-094622-esp12e/src/src.ino
@@ -1,28 +1,32 @@
 #include <EEPROM.h>
 #include <ESP8266WiFi.h>
 #include <ESP8266HTTPClient.h>
-#include "web.h"
-// piny
+#include "header.h"
+#include "define.h"
 
+
+// konstanty-piny
 const int WAKEUP_PIN1 = 5;
 const int WAKEUP_PIN2 = 13;
+// buzzer 14
 const int BUZZER_PIN = 14;
 const int LED_PIN = 12;
-const int ENPin = 2;      // the number of the Enable pin
+const int ENPin = 2;
+
+// globalni promnene
 bool BUTTONSTATE1 = false;
 bool BUTTONSTATE2 = false;
+
+
 WiFiClient wificlient;
 HTTPClient http;
 
-// feedback platformy
 void feedback_timer(int duration, int quantity) {
-    // Bzucak na oznameni stavu
+    // feedback platformy
     for (int i = 0; i < quantity; i++) {
         digitalWrite(BUZZER_PIN, HIGH);
-        digitalWrite(LED_PIN, HIGH);
         delay(duration);
         digitalWrite(BUZZER_PIN, LOW);
-        digitalWrite(LED_PIN, LOW);
         delay(100);
     }
 }
@@ -37,10 +41,11 @@ bool pin_pressed() {
 }
 
 void http_post(String HTTP_CONNECTION) {
+    // posle http request zasuvce
     if (http.begin(wificlient, HTTP_CONNECTION)) {
         int httpCode = http.POST(
                 (pin_pressed()) ? readEEPROM(HTTP_POS1, HTTP_LEN) : readEEPROM(HTTP_POS2,
-                                                                               HTTP_LEN));// nacte z pameti prislusnou JSON zpravu
+                                                                               HTTP_LEN));
         String payload = http.getString();
 
         if (payload.indexOf("errors") > 0 || payload == "") {
@@ -49,20 +54,25 @@ void http_post(String HTTP_CONNECTION) {
         Serial.println(payload);
         http.end();
     } else {
-        feedback_timer(200, 2);
+        feedback_timer(200, 3);
     }
 
 }
 
 void parsingIP() {
+    // najde prislusnou ip adresu
     int offset = (pin_pressed()) ? IP_POSA : IP_POSB;
     int count = countIP(offset);
-    for (int i = 0; i < count; i++) {
-        String HTTP_CONNECTION = "http://";
-        HTTP_CONNECTION += readIP(i, offset);
-        HTTP_CONNECTION += "/netio.json";
-        http_post(HTTP_CONNECTION);
-
+    if(count == 0){
+        feedback_timer(200, 3);
+        ESPSleep();
+    } else {
+        for (int i = 0; i < count; i++) {
+            String HTTP_CONNECTION = "http://";
+            HTTP_CONNECTION += readIP(i, offset);
+            HTTP_CONNECTION += "/netio.json";
+            http_post(HTTP_CONNECTION);
+        }
     }
 }
 
@@ -72,7 +82,7 @@ void wifi_setup() {
     String password = readEEPROM(PASSWORD_POS, PASSWORD_LEN);
     char *s = const_cast<char *>(ssid.c_str()); // prevede string na char*
     char *p = const_cast<char *>(password.c_str());
-    delay(400);
+    //delay(300);
     if (WiFi.status() != WL_CONNECTED) {
         WiFi.begin(s, p);
         for (int i = 0; i < 50; i++) { // kontroluje zda bylo pripojeno k wifi, po 5s se vypne esp
@@ -83,9 +93,8 @@ void wifi_setup() {
         }
         if (WiFi.status() != WL_CONNECTED) {// pokud neni pripojen ohlasi chybu
             feedback_timer(200, 2);
-            Serial.println("Nepripojeno");
             WiFi.disconnect();
-            setWiFiServer2();
+            ESPSleep();
         }
     }
 }
@@ -98,9 +107,8 @@ bool check_conf_mode() {
         setWiFiServer2();
         return true;
     } else if (BUTTONSTATE1 && BUTTONSTATE2) {
-        feedback_timer(400, 3);
-        // digitalWrite(ENPin, LOW);
-        ESP.deepSleep(0);
+        feedback_timer(200, 4);
+        ESPSleep();
         return false;
     } else
         return false;
@@ -111,58 +119,44 @@ void send_message() {
     wifi_setup();
     parsingIP();
 }
-//OTESTOVAT
-// void test(){
-//     for(int i=0; i<= IP_LEN; i++){
-//         if(i ==0 || i%16==0)
-//             EEPROM.write(i,'#');
-//         else
-//             EEPROM.write(i, '$');
-//         if(i-IP_LEN ==0 || i-IP_LEN%16==0)
-//             EEPROM.write(i,'#');
-//         else
-//             EEPROM.write(i + IP_LEN, '$');
-//     }
-//     EEPROM.commit();
-// }
 
 void setup_boot() {
+    // nastaveni po startu
     pinMode(ENPin, OUTPUT);
     digitalWrite(ENPin, HIGH);
     pinMode(WAKEUP_PIN1, INPUT_PULLUP);
     pinMode(WAKEUP_PIN2, INPUT_PULLUP);
     BUTTONSTATE1 = digitalRead(WAKEUP_PIN1);
     BUTTONSTATE2 = digitalRead(WAKEUP_PIN2);
-    Serial.begin(115200);
-    EEPROM.begin(800);
+    //Serial.begin(115200);
+    EEPROM.begin(1024);
     pinMode(BUZZER_PIN, OUTPUT);
     pinMode(LED_PIN, OUTPUT);
     Serial.println("Načtení pinu");
-    //test
-    //test();
-    //test
 }
+void ESPSleep() {
+    // digitalWrite(ENPin, LOW);
+    ESP.deepSleep(0);
 
-void debug() { // debug zpravy z pameti
-    Serial.println(readEEPROM(SSID_POS, 64));
-    Serial.println(readEEPROM(PASSWORD_POS, 64));
-    //Serial.println(readEEPROM(IP_POS1, 15));
-    Serial.println(readEEPROM(HTTP_POS1, 100));
-    //Serial.println(readEEPROM(IP_POS2, 15));
-    Serial.println(readEEPROM(HTTP_POS2, 100));
-    Serial.println(BUTTONSTATE1);
-    Serial.println(BUTTONSTATE2);
 }
+// void debug() { 
+//     // debug zpravy z pameti
+//     Serial.println(readEEPROM(SSID_POS, 64));
+//     Serial.println(readEEPROM(PASSWORD_POS, 64));
+//     Serial.println(readEEPROM(HTTP_POS1, 100));
+//     Serial.println(readEEPROM(HTTP_POS2, 100));
+//     Serial.println(BUTTONSTATE1);
+//     Serial.println(BUTTONSTATE2);
+// }
 
 void setup() {
     setup_boot();
-    debug();
+    // debug();
     if (!check_conf_mode()) {
         digitalWrite(LED_PIN, HIGH);
         Serial.println("nekonf");
         send_message();
-        // digitalWrite(ENPin, LOW);
-        ESP.deepSleep(0);
+        ESPSleep();
 
     }
 }
@@ -171,8 +165,8 @@ void loop() {
     delay(1);
     handleServer();
     if (!digitalRead(WAKEUP_PIN1) && !digitalRead(WAKEUP_PIN2)) {
-        //digitalWrite(ENPin, LOW);
-        ESP.deepSleep(0);
+        feedback_timer(400, 1);
+        ESPSleep();
     }
 
 }
