+++ kod/printed_updated/210715-094622-esp12e/src/web.cpp
@@ -4,57 +4,17 @@
 #include <EEPROM.h>
 
 // Vytvorene headery
-#include "web.h"
+#include "header.h"
 #include "HTML.h"
+#include "define.h"
 
 const int ENPin = 2; //deklarovano take v hlavnim programu
 const char *OWN_SSID = "NETIO_BUTTON";
-// Konstanty pro ulozeni do EEPROM
 
 bool APWork = true;
 
 ESP8266WebServer server(80);
 IPAddress OWN_IP(192, 168, 4, 1);
-
-// Prace s pameti
-
-void saveToEEPROM(String sToSave, int startPosition, int len) {
-    char arr[sToSave.length() + 1];
-    strcpy(arr, sToSave.c_str());
-    for (int i = 0; i < sToSave.length(); i++) {
-        EEPROM.write(startPosition + i, arr[i]);
-        if (i == len - 1)
-            break;
-    }
-    EEPROM.write(startPosition + sToSave.length(), -1);
-    EEPROM.commit();
-    delay(500);
-}
-
-String readEEPROM(int numberOfStart, int len) {
-    String tempNetio = "";
-    for (int i = 0; i < len; i++) {
-        char netio = EEPROM.read(numberOfStart + i);
-        if (netio == 255)
-            break;
-        tempNetio += netio;
-    }
-    return tempNetio;
-}
-
-
-//NOVE
-void saveToEEPROMIP(String sToSave, int startPosition) {
-    for (int i = 0 + startPosition; i < 160 + startPosition; i += 16) {
-        char isItIp = EEPROM.read(i);
-        if (isItIp == 255) {
-            EEPROM.write(i, '#');
-            saveToEEPROM(sToSave, i + 1, 15);
-            break;
-        }
-    }
-}
-
 
 // Funkce pro prevod
 
@@ -65,111 +25,12 @@ int dBmtoPercentage(int dBm) {
     return quality;
 }
 
-// prevod ip adresy na string
 String ipToString(IPAddress ip) {
+    // prevod ip adresy na string
     String s = "";
     for (int i = 0; i < 4; i++)
         s += i ? "." + String(ip[i]) : String(ip[i]);
     return s;
-}
-
-int countIP(int offset) {
-    int numberOfIP = 0;
-    for (int i = offset; i < 159 + offset; i += 16) {
-        char character = EEPROM.read(i);
-        if (character == '#')
-            numberOfIP++;
-    }
-    return numberOfIP;
-}
-
-String readIP(int i, int offset) {
-    int position = (i * 16) + offset;
-    if (EEPROM.read(position) == '#')
-        return readEEPROM(position + 1, 15);
-    else {
-        for (int j = position; position < IP_POSB + offset; position += 16) {
-            char character = EEPROM.read(position);
-            if (character == 35)
-                return readEEPROM(position + 1, 15);
-        }
-        return "?";
-    }
-}
-
-void deleteIP(int i, int offset) {
-    int position = (i * 16) + offset;
-    if (EEPROM.read(position) == '#') {
-        for (int j = 0; j < 16; j++)
-            EEPROM.write(j + position, 255);
-    } else {
-        for (int j = position; position < IP_POSB + offset; position += 16) {
-            char character = EEPROM.read(position);
-            if (character == 35) {
-                for (int k = 0; k < 16; k++)
-                    EEPROM.write(j + position, 255);
-            }
-        }
-    }
-    EEPROM.commit();
-}
-
-// Vytvoření JSON
-String jsonOfIP() {
-    int offset = IP_POSA;
-    int numberOfIP = countIP(offset);
-    String IPs = "{\"numOfIPA\": \"";
-    IPs += numberOfIP;
-    IPs += "\", \"IP_adressA\": [";
-    for (int i = 0; i < numberOfIP; i++) {
-        IPs += "\"";
-        IPs += readIP(i, offset);
-        IPs += "\"";
-        IPs += (i + 1 == numberOfIP) ? "" : ", ";
-    }
-    offset = IP_POSB;
-    numberOfIP = countIP(offset);
-    IPs += "], \"numOfIPB\": \"";
-    IPs += numberOfIP;
-    IPs += "\", \"IP_addressB\": [";
-    for (int i = 0; i < numberOfIP; i++) {
-        IPs += "\"";
-        IPs += readIP(i, offset);
-        IPs += "\"";
-        IPs += (i + 1 == numberOfIP) ? "" : ", ";
-    }
-    IPs += "]}";
-    Serial.println(IPs);
-    return IPs;
-}
-
-String jsonOfNetworks() {
-    int numberOfNetworks = WiFi.scanNetworks();
-    String networks = "{\"numOfNetworks\": \"";
-    networks += numberOfNetworks;
-    networks += "\", \"networks\": [";
-    for (int i = 0; i < numberOfNetworks; i++) {
-        networks += "\"";
-        networks += WiFi.SSID(i);
-        networks += "\"";
-        networks += (i + 1 == numberOfNetworks) ? " " : ", ";
-    }
-    networks += "], \"strengh\": [";
-    for (int i = 0; i < numberOfNetworks; i++) {
-        networks += "\"";
-        networks += dBmtoPercentage(WiFi.RSSI(i));
-        networks += " %\" ";
-        networks += (i + 1 == numberOfNetworks) ? " " : ", ";
-    }
-    networks += "], \"protection\": [";
-    for (int i = 0; i < numberOfNetworks; i++) {
-        networks += "\"";
-        networks += (WiFi.encryptionType(i) == 0) ? "None" : "Encrypted";
-        networks += "\" ";
-        networks += (i + 1 == numberOfNetworks) ? " " : ", ";
-    }
-    networks += "]}";
-    return networks;
 }
 
 // Nastavení AP a přípojení k WiFi
@@ -178,13 +39,13 @@ void connectToWiFi() {
     String password = readEEPROM(PASSWORD_POS, PASSWORD_LEN);
     char *s = const_cast<char *>(ssid.c_str());
     char *p = const_cast<char *>(password.c_str());
-    Serial.println(s);
+    // Serial.println(s);
     WiFi.begin(s, p);
+    delay(1000);
 }
 
 void APSet() {
     WiFi.mode(WIFI_AP_STA);
-    //WiFi.softAPConfig(OWN_IP, OWN_IP, IPAddress(255, 255, 255, 0));
     WiFi.softAP(OWN_SSID);
     delay(2000);
 }
@@ -220,7 +81,14 @@ void handleNetioAdd() {
 }
 
 void handleConfig() {
-    server.send(200, "text/html", configHTML);
+    String HTML = configHTML;
+    HTML += "<br> <h3>Button #1:</h3><br>";
+    HTML += readEEPROM(HTTP_POS1, HTTP_LEN);
+    HTML += "<br> <h3>Button #2:</h3><br>";
+    HTML += readEEPROM(HTTP_POS2, HTTP_LEN);
+    HTML += "</center></body></html>";
+
+    server.send(200, "text/html", HTML);
 }
 
 // Server - dynamické stránky
@@ -265,12 +133,12 @@ void handleWiFiPasswordRedirect() {
 
 void handleWiFiApprove() {
     if (WiFi.status() == WL_CONNECTED) {
-        String currentIP = "ESP IP: ";
+        String currentIP = "<center><h3>ESP IP: </h3><br>";
         currentIP += ipToString(WiFi.localIP());
-        currentIP += "<button style=\"background-color: #4CAF50;color: white;width: 200px;padding: 15px 32px;\" onClick=\"location.href = '/'\">Return</button>";
+        currentIP += "<br><button style=\"background-color: #005F41;color: white;width: 200px;padding: 15px 32px;\" onClick=\"location.href = '/'\">Return</button>";
         server.send(200, "text/html", currentIP);
     } else {
-        String badPassword = "Nesprávné heslo nebo chyba připojení <br><button onClick=\"location.href = '/wifi?ssid=";
+        String badPassword = "Password is not correct or device not able to connect <br><button onClick=\"location.href = '/wifi?ssid=";
         badPassword += readEEPROM(SSID_POS, SSID_LEN);
         badPassword += "';\">Return</button>";
         server.send(200, "text/html", badPassword);
@@ -279,11 +147,11 @@ void handleWiFiApprove() {
 
 void handleConfigCheck() {
     String html = "<meta http-equiv = \"refresh\" content = \"2; url = /buttonConfigure\" />";
-    if (server.hasArg("button1")) {
+    if (server.arg("button1")!= NULL) {
         saveToEEPROM(server.arg("button1"), HTTP_POS1, HTTP_LEN);
         server.send(200, "text/html", html);
     }
-    if (server.hasArg("button2")) {
+    if (server.arg("button2")!= NULL) {
         saveToEEPROM(server.arg("button2"), HTTP_POS2, HTTP_LEN);
         server.send(200, "text/html", html);
     }
@@ -291,28 +159,7 @@ void handleConfigCheck() {
 
 void handleDeepSleep() {
     feedback_timer(400, 1);
-    //digitalWrite(ENPin, LOW);
-    ESP.deepSleep(0);
-}
-
-void handleSettings() {
-    String html = "<head><style>body{font-family: \"Helvetica\";}</style></head><center><h1>Settings</h1><br>";
-    html += "<h4>IP address #1</h4>";
-    html += readEEPROM(IP_POSA, IP_LEN);
-    html += "<br>";
-    html += "<h4>IP address #2</h4>";
-    html += readEEPROM(IP_POSB, IP_LEN);
-    html += "<br>";
-    html += "<h4>JSON #1</h4>";
-    html += readEEPROM(HTTP_POS1, IP_LEN);
-    html += "<br>";
-    html += "<h4>JSON #2</h4>";
-    html += readEEPROM(HTTP_POS2, IP_LEN);
-    html += "<br>";
-    html += "<h4>SSID</h4>";
-    html += readEEPROM(SSID_POS, SSID_LEN);
-    html += "<br></center>";
-    server.send(200, "text/html", html);
+    ESPSleep();
 }
 
 void handleDisconnect() {
@@ -324,19 +171,19 @@ void handleDisconnect() {
     server.send(200, "text/html", "<meta http-equiv = \"refresh\" content = \"2; url = /\" />");
 }
 
-void handledebug() {
-    String pes = "test: ";
-    for (int i = 0; i < 400; i++) {
-        if (i == 188)
-            pes += "---";
-        else if (EEPROM.read(i) == 255)
-            pes += "+";
-        else
-            pes += EEPROM.read(i);
-        pes += ";";
-    }
-    server.send(200, "text/html", pes);
-}
+// void handledebug() {
+//     String pes = "test: ";
+//     for (int i = 0; i < 400; i++) {
+//         if (i == 188)
+//             pes += "---";
+//         else if (EEPROM.read(i) == 255)
+//             pes += "+";
+//         else
+//             pes += EEPROM.read(i);
+//         pes += ";";
+//     }
+//     server.send(200, "text/html", pes);
+// }
 
 void serversOn() {
     server.on("/scannedWiFi.json", handleScanWiFi);
@@ -351,29 +198,16 @@ void serversOn() {
     server.on("/netioProduct/check", HTTP_GET, handleNetioDevice);
     server.on("/buttonConfigure", handleConfig);
     server.on("/buttonConfigure/check", HTTP_POST, handleConfigCheck);
-    server.on("/settings", handleSettings);
     server.on("/deepsleep", handleDeepSleep);
     server.on("/disconnect", handleDisconnect);
-    server.on("/debug", handledebug);
+    // server.on("/debug", handledebug);
     server.begin();
-
-
-}
-
-void setWiFiServer() {
-    APSet();
-    delay(500);
-    serversOn();
 }
 
 void setWiFiServer2() {
-    connectToWiFi();
-    delay(1000);
-    if (WiFi.status() != WL_CONNECTED) {
-        feedback_timer(200, 2);
-        setWiFiServer();
-    } else
-        serversOn();
+    APSet();
+    delay(500);
+    serversOn();
 }
 
 void handleServer() {
